+++ app/src/main/java/ru/yandex/cup/live/images/ui/UiLayer.kt
@@ -1,6 +1,6 @@
 package ru.yandex.cup.live.images.ui
 
 data class UiLayer(
-    val index: Int,
+    val index: Long,
     val drawingQueue: List<UiPath>,
 )
